+++ Render/D3D12/D3D12Renderer.h
@@ -38,11 +38,11 @@ namespace Lightning
 				const RectI* rects=nullptr, std::size_t rectCount = 0)override;
 			void INTERFACECALL ClearDepthStencilBuffer(IDepthStencilBuffer* buffer, DepthStencilClearFlags flags, float depth, std::uint8_t stencil, 
 				const RectI* rects = nullptr, std::size_t rectCount = 0)override;
-			void INTERFACECALL ApplyRenderTargets(const IRenderTarget*const * renderTargets, std::uint8_t renderTargetCount, IDepthStencilBuffer* dsBuffer)override;
+			void INTERFACECALL ApplyRenderTargets(const IRenderTarget*const * renderTargets, std::size_t renderTargetCount, IDepthStencilBuffer* dsBuffer)override;
 			void INTERFACECALL ApplyPipelineState(const PipelineState& state)override;
 			void INTERFACECALL ApplyViewports(const Viewport* viewports, std::uint8_t viewportCount)override;
 			void INTERFACECALL ApplyScissorRects(const ScissorRect* scissorRects, std::uint8_t scissorRectCount)override;
-			void INTERFACECALL BindGPUBuffer(std::uint8_t slot, IGPUBuffer* buffer)override;
+			void INTERFACECALL BindGPUBuffer(std::size_t slot, IGPUBuffer* buffer)override;
 			void INTERFACECALL Draw(const DrawParam& param)override;
 			ID3D12CommandQueue* GetCommandQueue();
 			ID3D12GraphicsCommandList* GetGraphicsCommandList();
